benchmarks: call `valid_encoding?` on large strings
This is to make the comparison a bit more fair.

The `redis` gem simply call `force_encoding(Encoding.default_external)`
and doesn't bother to check if that string is valid.

`redis-client` however will apply the encoding and then check
if the string is valid, if it isn't it will mark the string as binary.

This difference makes the `redis` gem faster on benchmark as
it saves having to scan the string. However whenever that string
will operated on or concatenated, MRI will scan it.

So in a way `redis-client` pays the cost upfront, while `redis` defers
it.

So I believe that it's more fair to call `valid_encoding?` to trigger
that scan if it didn't happen yet.