Automatically reconnect if the process was forked
Initially I didn't want that feature. But I think there's just
way too much code out there that rely on this behavior that was
in redis-rb. Also it's not always easy to close connections from
an after_fork callback, even though it is the cleaner thing to do.

Ultimately the overhead is relatively minimal, so might as well
make it easy for users.