Fix for Xcode 14 compilation issues
As described in https://bugs.ruby-lang.org/issues/19005, Ruby
interpreters with XCode 14.0 will default without the
`-Wl,undefined,dynamic_lookup` compilation flag enabled. As a result,
any symbols that are not exported by the shared library will throw an
error during linking.

hiredis-client was failing to build on these platforms because
`_ruby_abi_version` is not exported in Ruby interpreters prior to 3.2.

To fix this, we need to export this symbol only if the Ruby version is
>= 3.2.0.

Closes #58