Implement Write Barriers for HiredisConnection
We use a bit of a hack here, for two reasons:

First we'd need a reference to the connection from the parsing functions,
it's possible but a bit annoying.

Also if we were to fire write barriers ourselves from the connection object,
we'd be very likely to promote the response objects to the old gen, which
isn't desirable.

So instead we allocate an RArray and keep a reference to it on the stack.
That array is used to store the elements of the response being parsed.