Draft: Replace bypass BufferedIO when `IO#timeout` is available
Ruby IOs have an internal buffer already, the only reason we have
to buffer again ourselves is to be able to timeout on `IO#gets`.

With `IO#timeout` being introduced in Ruby 3.2, we can get rid
of the second buffer.

This is just a PoC, there is a handful of test failures, I just
wanted to explore.