Fix a GC bug in hiredis-client
Since we only mark response values through the `stack` Array,
these values are movable, so we can't trust the reply
retuned by `redisGetReplyFromReader` as it may be outdated if
the value was moved.

Instead we can directly take the first element of `stack` as it
should be out response object.