Add unit test/comment for multi/exec handling
In a mulit/exec command, the results of `call_pipelined` can
be an array. It was not obvious to me this was handled outside
of the `ConnectionMixin`. This commit adds a unit test to ensure
the commands for an error are stored properly in this case and a
comment mentioning that the results are handled in `Multi#_coerce!`.