Ensure BLPOP/BRPOP returns nil instead of raising ReadTimeoutError
In https://github.com/redis/redis-rb/issues/1279, we discovered that
`redis-rb` properly returned `nil` when `timeout` was reached with no
key present, but when connecting to Redis Sentinels, the client raised
a `ReadTimeoutTimeout` error.

This occurred because of a subtle difference in how `RedisClient`
(from `redis-rb`) and `Redis::Client` (from `redis-client`) behaved.
The former, which is used with standalone Redis, returned `nil`
because the socket read timeout was incremented to the command timeout
value (https://github.com/redis/redis-rb/pull/1175). The latter did
not have this, so the socket read timeout would get triggered before
the actual Redis timeout hit.

To make the behavior consistent, increment the configured read timeout
to the command timeout.

Closes https://github.com/redis/redis-rb/issues/1279